게시글 테스트 코드 및 리팩터링

+++ src/main/java/com/example/spotserver/domain/PageInfo.java
@@ -7,7 +7,7 @@ public class PageInfo {
 
     private int page;
     private int size;
-    private int numberOfElements;
+    private Long numberOfElements;
     private Long totalElements;
     private int totalPage;
